feat: Show splashscreen when application goes to background
We wan't the app to display the SplashScreen when sent to background

When done from react-native side, the app may not have time to
correctly display the SplashScreen before being frozen by the OS, so we
need to do it from the native side

On iOS the `UIApplicationWillResignActiveNotification` event is used so
the SplashScreen can be displayed when the user starts to use the
multi-task feature. We cannot use
`UIApplicationDidEnterBackgroundNotification` as it occurs too late and
the UI won't refresh in time

On Android the `onHostPause` event is used and corresponds to when the
app is sent to background. When the user starts to use the multi-task
feature nothing happens, this event only occurs when the user goes to
the device's home or when switching app

Only displaying the SplashScreen is done on native side, the action of
hiding it should be done in react-native side so we have better control
on when to do it